# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для CDP сервиса",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API"
        }
    ],
    "paths": {
...
        "api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "registerUser",
                "parameters": [
                    {
                        "name": "mall_id",
                        "in": "path",
                        "description": "id ТРЦ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    },
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Номер телефона пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "89853429241"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Имя пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Вячеслав"
                    },
                    {
                        "name": "surname",
                        "in": "path",
                        "description": "Фамилия пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Погольша"
                    },
                    {
                        "name": "sex",
                        "in": "path",
                        "description": "Пол пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "male,female"
                    },
                    {
                        "name": "birthdate",
                        "in": "path",
                        "description": "День рождения пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "04/05/2002"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "solvintech@mail.ru"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "name": "mall_id",
                        "in": "path",
                        "description": "id ТРЦ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    },
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Номер телефона пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "89853429321"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "bearerToken": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/verify": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Отправка смс-кода для подтверждения телефона",
                "description": "Отправка смс-кода",
                "operationId": "registerVerifyPhone",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Номер телефона пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "89853429213"
                    },
                    {
                        "name": "mall_id",
                        "in": "path",
                        "description": "id ТРЦ пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}